🚚 Move Subscription class to deprecated.php
This class has been replaced by the classes in the
dokuwiki/Subscriptions namespace.
